fix pissa and olora
(cherry picked from commit cadf75ea3c42a628e4e56efd4fd11c2a348b58dc)
