+++ swift/llm/tuner.py
@@ -13,6 +13,7 @@ from swift.tuners import (AdaLoraConfig, AdapterConfig, BOFTConfig, IA3Config, L
 from swift.utils import activate_model_parameters, freeze_model_parameters, get_logger, use_torchacc
 from swift.utils.module_mapping import MODEL_KEYS_MAPPING
 from .utils import SftArguments, find_all_linears, find_embedding, find_ln, is_adapter
+from .utils.callbacks import DynamicLayerActivationCallback, TrainerAdapterCallback, ConvertWeightsCallback
 
 logger = get_logger()
 
@@ -321,52 +322,6 @@ def prepare_model(model, args: SftArguments):
     callbacks = []
     if args.lisa_activated_layers > 0:
         assert args.sft_type == 'full', 'LISA only supports full parameter training.'
-
-        class DynamicLayerActivationCallback(TrainerCallback):
-
-            def __init__(self, n_layers: int, step_interval: int, model: torch.nn.Module):
-                super().__init__()
-                self.n_layers = n_layers
-                self.step_interval = step_interval
-                self.model = model
-                layers_name = None
-                layers = None
-                for name, module in model.named_modules():
-                    if isinstance(module, torch.nn.ModuleList):
-                        layers_name = name
-                        layers = module
-                        break
-                assert layers_name is not None
-                self.layers_attribute = layers_name
-                self.total_layers = len(layers)
-
-                # Freeze all layers upon initialization
-                self.freeze_all_layers()
-                self.active_layers_indices = []
-
-            def freeze_all_layers(self):
-                layers = self.model.get_submodule(self.layers_attribute)
-                for layer in layers:
-                    for param in layer.parameters():
-                        param.requires_grad = False
-
-            def on_step_begin(self, args, state, control, **kwargs):
-                # Check if it's time to switch active layers, including at step 0
-                if state.global_step % self.step_interval == 0 or state.global_step == 1:
-                    self.switch_active_layers()
-
-            def switch_active_layers(self):
-                # First, disable gradients for all layers
-                self.freeze_all_layers()
-
-                # Randomly select n_layers to activate
-                layers = self.model.get_submodule(self.layers_attribute)
-                self.active_layers_indices = np.random.choice(range(self.total_layers), self.n_layers, replace=False)
-                # Enable gradients only for the selected layers
-                for idx in self.active_layers_indices:
-                    for param in layers[idx].parameters():
-                        param.requires_grad = True
-
         lisa_callback = DynamicLayerActivationCallback(
             n_layers=args.lisa_activated_layers,  # Number of layers to activate
             step_interval=args.lisa_step_interval,  # Step interval to update active layers
@@ -374,29 +329,9 @@ def prepare_model(model, args: SftArguments):
         lisa_callback.switch_active_layers()  # Make trainable parameters printing a correct value
         callbacks.append(lisa_callback)
 
-    class TrainerAdapterCallback(TrainerCallback):
-
-        def __init__(self):
-            self.global_step = 0
-
-        # offload original_modules to cpu, to save memory
-        def on_train_begin(self, _args, state, control, **kwargs):
-            if hasattr(model, 'set_active_adapters'):
-                model.set_active_adapters(model.adapters.keys(), offload='cpu')
-            if args.sft_type == 'adalora':
-                model.peft_config['default'].total_step = state.max_steps
-
-                def zero_grad(_self, *args, **kwargs):
-                    _self.update_and_allocate(self.global_step + 1)
-                    _self._zero_grad(*args, **kwargs)
-
-                model._zero_grad = model.zero_grad
-                model.zero_grad = types.MethodType(zero_grad, model)
-
-        def on_step_end(self, _args, state, control, **kwargs):
-            if args.sft_type == 'adalora':
-                self.global_step = state.global_step
+    if 'pissa' in args.init_lora_weights or 'olora' in args.init_lora_weights:
+        callbacks.append(ConvertWeightsCallback())
 
     if is_adapter(args.sft_type) and args.tuner_backend == 'swift':
-        callbacks.append(TrainerAdapterCallback())
+        callbacks.append(TrainerAdapterCallback(args))
     return model, callbacks
